Change ChallengeDuration to uint32
Per https://github.com/statechannels/go-nitro/pull/877#issuecomment-1251235127

+++ client/engine/chainservice/adjudicator/challenge_test.go
@@ -152,7 +152,7 @@ func TestChallenge(t *testing.T) {
 		}
 
 		// Generate expectation
-		expectedFinalizesAt := header.Time + s.ChallengeDuration
+		expectedFinalizesAt := header.Time + uint64(s.ChallengeDuration)
 		cId := s.ChannelId()
 		expectedOnChainStatus, err := generateStatus(s, expectedFinalizesAt)
 		if err != nil {
